+++ src/NamespacesConsumer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { I18nContext, withI18n } from './context';
-import { deprecated } from './utils';
+import { warnOnce, deprecated, initSSR } from './utils';
 
 let removedIsInitialSSR = false;
 
@@ -8,23 +8,29 @@ export class NamespacesConsumerComponent extends Component {
   constructor(props) {
     super(props);
 
+    if (!props.i18n) {
+      return warnOnce(
+        'You will need pass in an i18next instance either by props, using I18nextProvider or by using i18nextReactModule. Learn more https://react.i18next.com/components/overview#getting-the-i-18-n-function-into-the-flow'
+      );
+    }
+
     // nextjs / SSR: getting data from next.js or other ssr stack
-    if (props.initialI18nStore) {
-      props.i18n.services.resourceStore.data = props.initialI18nStore;
-      props.i18nOptions.wait = false; // we got all passed down already
-    }
-    if (props.initialLanguage) {
-      props.i18n.changeLanguage(props.initialLanguage);
-    }
+    initSSR(props);
 
     // provider SSR: data was set in provider and ssr flag was set
     if (props.i18n.options && props.i18n.options.isInitialSSR) {
       props.i18nOptions.wait = false;
     }
 
+    // reportNS if needed for SSR
+    const namespaces = this.getNamespaces();
+    if (props.reportNS) {
+      namespaces.forEach(props.reportNS);
+    }
+
+    // check if we could flag this ready already as all is loaded
     const language = props.i18n.languages && props.i18n.languages[0];
-    const ready =
-      !!language && this.getNamespaces().every(ns => props.i18n.hasResourceBundle(language, ns));
+    const ready = !!language && namespaces.every(ns => props.i18n.hasResourceBundle(language, ns));
 
     this.state = {
       i18nLoadedAt: null,
@@ -80,15 +86,20 @@ export class NamespacesConsumerComponent extends Component {
       i18nOptions.nsMode === 'fallback'
         ? namespaces
         : namespaces && namespaces.length
-          ? this.getNamespaces()[0]
+          ? namespaces[0]
           : 'translation'
     );
   }
 
   getNamespaces() {
-    const { i18n, ns } = this.props;
-    const namespace = ns || (i18n.options && i18n.options.defaultNS);
-    return typeof namespace === 'string' ? [namespace] : namespace;
+    const { i18n, ns, defaultNS } = this.props;
+
+    const namespaces =
+      typeof ns === 'function'
+        ? ns(this.props)
+        : ns || defaultNS || (i18n.options && i18n.options.defaultNS);
+
+    return typeof namespaces === 'string' ? [namespaces] : namespaces || [];
   }
 
   loadNamespaces() {
@@ -162,13 +173,9 @@ export class NamespacesConsumerComponent extends Component {
 
 export const NamespacesConsumer = withI18n()(NamespacesConsumerComponent);
 
-let warnedI18n;
 export function I18n(props) {
-  if (!warnedI18n) {
-    deprecated(
-      'I18n was renamed to "NamespacesConsumer" to make it more clear what the render prop does.'
-    );
-    warnedI18n = true;
-  }
+  deprecated(
+    'I18n was renamed to "NamespacesConsumer" to make it more clear what the render prop does.'
+  );
   return <NamespacesConsumer {...props} />;
 }