+++ src/withNamespaces.js
@@ -15,19 +15,6 @@ export function withNamespaces(namespaceArg, options = {}) {
       constructor(props) {
         super(props);
 
-        this.namespaces =
-          typeof namespaceArg === 'function'
-            ? namespaceArg(props)
-            : namespaceArg ||
-              props.defaultNS ||
-              (props.i18n.options && props.i18n.options.defaultNS);
-        if (typeof this.namespaces === 'string') this.namespaces = [this.namespaces];
-
-        if (props.reportNS) {
-          const namespaces = this.namespaces || [undefined];
-          namespaces.forEach(props.reportNS);
-        }
-
         this.getWrappedInstance = this.getWrappedInstance.bind(this);
       }
 
@@ -55,7 +42,7 @@ export function withNamespaces(namespaceArg, options = {}) {
       }
 
       render() {
-        const { i18nOptions } = this.props;
+        const { namespaces, i18nOptions } = this.props;
         const mergedI18nOptions = { ...i18nOptions, ...options };
         const extraProps = {};
 
@@ -67,7 +54,11 @@ export function withNamespaces(namespaceArg, options = {}) {
 
         return React.createElement(
           NamespacesConsumer,
-          { ns: this.namespaces, ...this.props, i18nOptions: mergedI18nOptions },
+          {
+            ns: namespaces || namespaceArg,
+            ...this.props,
+            i18nOptions: Object.keys(mergedI18nOptions).length > 0 ? mergedI18nOptions : null,
+          },
           (t, { ready, ...rest }) =>
             React.createElement(WrappedComponent, {
               tReady: ready,
@@ -92,13 +83,7 @@ export function withNamespaces(namespaceArg, options = {}) {
 withNamespaces.setDefaults = setDefaults;
 withNamespaces.setI18n = setI18n;
 
-let warnedTranslate;
 export function translate(ns, opts) {
-  if (!warnedTranslate) {
-    deprecated(
-      'translate was renamed to "withNamespaces" to make it more clear what the HOC does.'
-    );
-    warnedTranslate = true;
-  }
+  deprecated('translate was renamed to "withNamespaces" to make it more clear what the HOC does.');
   return withNamespaces(ns, opts);
 }