+++ test/withNamespaces.spec.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { shallow, render, mount } from 'enzyme';
 import { withNamespaces } from '../src/withNamespaces';
 
@@ -68,7 +67,7 @@ describe('withNamespaces', () => {
     const namespaces = [];
     const C = withNamespaces()(() => <div>text</div>);
     mount(<C reportNS={ns => namespaces.push(ns)} />);
-    expect(namespaces).toEqual([undefined]);
+    expect(namespaces).toEqual([]);
   });
   it('should report default namespace if no namespace used', () => {
     const i18n = {